Split explain utility, invariance tests.
